Remove util.java since we moved the logic parts into TagService.
